Split ann_benchmark.go into focused files
The original file was ~1274 lines handling schema creation, compression,
multi-tenancy, async indexing, query execution, and CLI wiring all in one
place. Split into four files with clear single responsibilities:

- ann_schema.go  — schema CRUD, updateEf, tenant management, deletes
- ann_ingest.go  — CompressionType/Batch types, createClient, writeChunk,
                   waitReady, enableCompression
- ann_query.go   — parseEfValues, runQueries, shouldStopRunQueries,
                   benchmarkANN/Duration, median/sampledResults
- ann_benchmark.go — ResultsJSONBenchmark, uuidFromInt/intFromUUID,
                     cobra command + flag registration only

All existing unit tests and integration tests continue to pass.
